+++ lib/src/components/components.dart
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 export 'flowbite_alert.dart';
+export 'flowbite_badge.dart';
 export 'flowbite_button.dart';
 export 'flowbite_icon_shape.dart';
 export 'flowbite_input_field.dart';